Move sound enums into Engine/Data
Sound ids and music ids are game data, and SoundType/SoundFlags are
fields of a dsounds.bin row (see SoundInfo), so the header belongs with
the other game data enums rather than next to AudioPlayer. Living up in
Media/Audio meant the bottom-most data package had to reach up to the
audio layer: Engine/Data/TileEnumFunctions.h needs SoundId for
walkSoundForTileset, and data structs in Engine, Engine/Objects and
Engine/Evt embed SoundId fields.

This removes every header-level dependency the engine packages had on
Media/Audio.
