+++ src/Engine/Party.h
@@ -12,7 +12,7 @@
 #include "Engine/Time/Time.h"
 #include "Engine/Time/Timer.h"
 
-#include "Media/Audio/SoundEnums.h"
+#include "Engine/Data/SoundEnums.h"
 
 #include "Library/Geometry/Point.h"
